requirements.txt
streamlit==1.14.0
streamlit-extras==0.2.0
deta==1.1.0
streamlit-nested-layout==0.1.1
deta-base-pydantic==0.1.1